#2 Started work on more movement mechanics
So far I have a static speedline image that slowly appears as you gain speed but I would like a gif in the future. I also made an acceleration so it slowly transitions to the max speed when running. Right now I think i need to make a deceleration as well or come up with a different system for speeding up and down.